move QabElem <-> GAP cyclotomics conversion
Now the conversion (both directions) is in `src/GAP/gap_to_oscar.jl` and `src/GAP/oscar_to_gap.jl`, as suggested in the discussion of PR #840.

For that, the ordering of `include` statements in `src/Oscar.jl` had to be changed.
