+++ test/GAP/oscar_to_gap.jl
@@ -79,10 +79,16 @@ end
 end
 
 @testset "single cyclotomics" begin
+    # from cyclotomic fields
     F, z = CyclotomicField(5)
     e5 = GAP.Globals.E(5)
     @test GAP.Obj(z^2+z+1) == e5^2 + e5 + 1
 
+    # from `QabElem`
+    F, z = abelian_closure(QQ)
+    @test GAP.Obj(z(5) + z(5)^4) == e5 + e5^4
+
+    # not supported conversions
     F, z = quadratic_field(5)
     @test_throws ArgumentError GAP.Obj(z)
 end
